clash-vision/requirements.txt
ultralytics==8.2.103
opencv-python
numpy
albumentations
pyyaml
mss
rich
typer
Pillow
